move all nonce interactions into userpasswd
This feels correct, since the bit that generates the URLs now lives with
the code that consumes it
